refactor: rename `Nazuki.VM.Arch2562` to `Nazuki.VM.Word1` (not sure)
